refactor: ♻️ re-organize lib directory
🏁 Closes: #505
