+++ src/components/media/crew-list.tsx
@@ -1,6 +1,6 @@
 import { Link } from "@tanstack/react-router";
 
-import { tmdbImageUrl } from "@/lib/urls";
+import { tmdbImageUrl } from "@/lib/tmdb-images";
 
 import { ListContent } from "../shared/list-content";
 import { CardImageFallback } from "./card-image-fallback";
