+++ src/components/movie/movie-details.tsx
@@ -1,8 +1,8 @@
 import { Link, Outlet } from "@tanstack/react-router";
 import { Suspense } from "react";
 
-import { asQuote } from "@/lib/as-quote";
-import { tmdbImageUrl } from "@/lib/urls";
+import { quote } from "@/lib/string";
+import { tmdbImageUrl } from "@/lib/tmdb-images";
 
 import { MediaGenres } from "../media/media-genres";
 import { MediaRating } from "../media/media-rating";
@@ -55,7 +55,7 @@ export const MovieDetails = ({ movie }: MovieDetailsProps) => {
             </div>
             <Prose size="lg">
               <h1>{movie.title}</h1>
-              {movie.tagline ? <p>{asQuote(movie.tagline)}</p> : null}
+              {movie.tagline ? <p>{quote(movie.tagline)}</p> : null}
               <p>{movie.overview}</p>
             </Prose>
             <div className="flex justify-center gap-2 md:justify-start">
