+++ src/components/people/person-card.tsx
@@ -1,6 +1,6 @@
 import { Link } from "@tanstack/react-router";
 
-import { tmdbImageUrl } from "@/lib/urls";
+import { tmdbImageUrl } from "@/lib/tmdb-images";
 
 import { CardImageFallback } from "../media/card-image-fallback";
 import { MediaType } from "../media/media-type";
