+++ src/lib/date.ts
@@ -17,3 +17,7 @@ const dateFormatter = new Intl.DateTimeFormat(undefined, {
 export const date = (date: string): string => {
   return dateFormatter.format(new Date(date));
 };
+
+export const year = (date: string) => {
+  return new Date(date).getFullYear();
+};
