--- src/lib/types.ts
@@ -1,1 +1,0 @@
-export type MediaType = "movie" | "person" | "tv";
