--- src/lib/year.ts
@@ -1,3 +1,0 @@
-export const year = (date: string) => {
-  return new Date(date).getFullYear();
-};
