+++ src/routes/_layout/movies.$id.trailer.tsx
@@ -3,7 +3,7 @@ import { createFileRoute } from "@tanstack/react-router";
 
 import { movieVideosOptions } from "@/api/movie/details.videos";
 import { MediaTrailerDialog } from "@/components/media/media-trailer-dialog";
-import { selectYoutubeTrailer } from "@/lib/youtube";
+import { selectYoutubeTrailer } from "@/lib/tmdb-youtube";
 
 export const Route = createFileRoute("/_layout/movies/$id/trailer")({
   component: RouteComponent,
