+++ src/routes/_layout/movies.$id.tsx
@@ -7,7 +7,7 @@ import { movieExternalOptions } from "@/api/movie/details.external";
 import { MovieDetails } from "@/components/movie/movie-details";
 import { site } from "@/config/site";
 import { seo } from "@/lib/seo";
-import { tmdbImageUrl } from "@/lib/urls";
+import { tmdbImageUrl } from "@/lib/tmdb-images";
 import { PathParamsSchema } from "@/schemas/path-params";
 
 export const Route = createFileRoute("/_layout/movies/$id")({
