+++ src/routes/_layout/people.$id.tsx
@@ -7,7 +7,7 @@ import { personExternalOptions } from "@/api/people/details.external";
 import { PersonDetails } from "@/components/people/person-details";
 import { site } from "@/config/site";
 import { seo } from "@/lib/seo";
-import { tmdbImageUrl } from "@/lib/urls";
+import { tmdbImageUrl } from "@/lib/tmdb-images";
 import { PathParamsSchema } from "@/schemas/path-params";
 
 export const Route = createFileRoute("/_layout/people/$id")({
